Moved TextNormalization to core
