+++ LTU.SearchEngine.Backend/LTU.SearchEngine.Test/Core.Tests/TextNormalization/TextNormalizerTest.cs
@@ -1,9 +1,9 @@
-﻿using LTU.SearchEngine.Infrastructure.Indexing.Normalization;
+﻿using LTU.SearchEngine.Backend.Core.TextNormalization;
 using Moq;
 using System;
 using Xunit;
 
-namespace LTU.SearchEngine.Test.Indexing.Tests
+namespace LTU.SearchEngine.Test.Core.Tests.TextNormalization
 {
     public class TextNormalizerTests
     {
